Add labs path with layout distinct from GCN
