+++ app/root.tsx
@@ -21,36 +21,25 @@
 import { cssBundleHref } from '@remix-run/css-bundle'
 import type { DataFunctionArgs, LinksFunction } from '@remix-run/node'
 import {
-  Link,
   Links,
   LiveReload,
   Meta,
   Outlet,
   Scripts,
   ScrollRestoration,
-  isRouteErrorResponse,
   useLocation,
   useNavigation,
-  useRouteError,
   useRouteLoaderData,
 } from '@remix-run/react'
-import {
-  ButtonGroup,
-  FormGroup,
-  GovBanner,
-  GridContainer,
-} from '@trussworks/react-uswds'
+import { GovBanner } from '@trussworks/react-uswds'
 import type { ReactNode } from 'react'
 import TopBarProgress from 'react-topbar-progress-indicator'
 import { useSpinDelay } from 'spin-delay'
 import invariant from 'tiny-invariant'
 
 import { DevBanner } from './components/DevBanner'
-import { Footer } from './components/Footer'
-import { Header } from './components/Header'
-import NewsBanner from './components/NewsBanner'
 import { features, getEnvOrDieInProduction, origin } from './lib/env.server'
-import { type BreadcrumbHandle, Title } from './root/Title'
+import { Title } from './root/Title'
 import { getUser } from './routes/_gcn._auth/user.server'
 
 import highlightStyle from 'highlight.js/styles/github.css'
@@ -89,10 +78,6 @@ TopBarProgress.config({
     '1.0': '#0050d8',
   },
 })
-
-export const handle: BreadcrumbHandle = {
-  breadcrumb: 'GCN',
-}
 
 export const links: LinksFunction = () => [
   {
@@ -194,7 +179,7 @@ function Progress() {
   return <>{showProgress && <TopBarProgress />}</>
 }
 
-function Document({ children }: { children?: React.ReactNode }) {
+export default function () {
   const noIndex = useHostname() !== 'gcn.nasa.gov'
 
   return (
@@ -214,18 +199,7 @@ function Document({ children }: { children?: React.ReactNode }) {
         <Progress />
         <GovBanner />
         <DevBanner />
-        <Header />
-        <NewsBanner>
-          New Swift-BAT/GUANO and IceCube Notice Types Available! See{' '}
-          <Link
-            className="usa-link"
-            to="/news#new-swift-bat/guano-and-icecube-notices-available-via-kafka"
-          >
-            news and announcements
-          </Link>
-        </NewsBanner>
-        <main id="main-content">{children}</main>
-        <Footer />
+        <Outlet />
         <ScrollRestoration />
         <Scripts />
         <LiveReload />
@@ -233,104 +207,5 @@ function Document({ children }: { children?: React.ReactNode }) {
         <link rel="stylesheet" href={highlightStyle} />
       </body>
     </html>
-  )
-}
-
-function ErrorUnexpected({ children }: { children?: ReactNode }) {
-  return (
-    <Document>
-      <GridContainer className="usa-section">
-        <h1>Unexpected error {children}</h1>
-        <p className="usa-intro">An unexpected error occurred.</p>
-        <FormGroup>
-          <ButtonGroup>
-            <Link to="/" className="usa-button">
-              Go home
-            </Link>
-          </ButtonGroup>
-        </FormGroup>
-      </GridContainer>
-    </Document>
-  )
-}
-
-function ErrorUnauthorized() {
-  const url = useUrl()
-  return (
-    <Document>
-      <GridContainer className="usa-section">
-        <h1>Unauthorized</h1>
-        <p className="usa-intro">
-          We're sorry, you must log in to access the page you're looking for.
-        </p>
-        <p className="usa-paragraph">Log in to access that page, or go home.</p>
-        <FormGroup>
-          <ButtonGroup>
-            <Link
-              to={`/login?redirect=${encodeURIComponent(url)}`}
-              className="usa-button"
-            >
-              Log in and take me there
-            </Link>
-            <Link to="/" className="usa-button usa-button--outline">
-              Go home
-            </Link>
-          </ButtonGroup>
-        </FormGroup>
-      </GridContainer>
-    </Document>
-  )
-}
-
-function ErrorNotFound() {
-  return (
-    <Document>
-      <GridContainer className="usa-section">
-        <h1>Page not found</h1>
-        <p className="usa-intro">
-          We're sorry, we can't find the page you're looking for. It might have
-          been removed, changed its name, or is otherwise unavailable.
-        </p>
-        <p className="usa-paragraph">
-          Visit our homepage for helpful tools and resources, or contact us and
-          we'll point you in the right direction.
-        </p>
-        <FormGroup>
-          <ButtonGroup>
-            <Link to="/" className="usa-button">
-              Visit homepage
-            </Link>
-            <Link to="/contact" className="usa-button usa-button--outline">
-              Contact us
-            </Link>
-          </ButtonGroup>
-        </FormGroup>
-      </GridContainer>
-    </Document>
-  )
-}
-
-export function ErrorBoundary() {
-  const error = useRouteError()
-  if (isRouteErrorResponse(error)) {
-    switch (error.status) {
-      case 403:
-        return <ErrorUnauthorized />
-      case 404:
-        return <ErrorNotFound />
-      default:
-        return <ErrorUnexpected>HTTP {error.status}</ErrorUnexpected>
-    }
-  } else {
-    console.error(error)
-    return <ErrorUnexpected />
-  }
-}
-
-export default function App() {
-  return (
-    <Document>
-      <Outlet />
-    </Document>
   )
 }
