+++ app/routes/labs._index.tsx
@@ -1,0 +1,10 @@
+/*!
+ * Copyright © 2023 United States Government as represented by the
+ * Administrator of the National Aeronautics and Space Administration.
+ * All Rights Reserved.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+export default function () {
+  return <p>Hello world</p>
+}
